+++ bin/TA_dmarc_rh_dmarc_imap.py
@@ -57,6 +57,13 @@ fields = [
         default=True,
         validator=None
     ), 
+    field.RestField(
+        'validate_xml',
+        required=False,
+        encrypted=False,
+        default=True,
+        validator=None
+    ), 
 
     field.RestField(
         'disabled',